Pre-fill contact information in PayPal Guest Checkout (#477)
* Send phone number from order to PayPal

* Send email address from order to PayPal

* Send email address from customer to PayPal, when initiating flow prior to order creation

* Set and retrieve phone and email on WC()->customer, for pre-filling with SPB

Adds phone and email to the set of values set on the customer during Smart Payment Button initialization, and makes sure phone number is retrieved from customer object to be passed to PayPal. Also slight refactor moving the 'country' lines up to match order of fields.

* Access order and customer properties directly for WC <3.0 case

Supports phone, email, first name, and last name even when accessors are not available on the object.