Add hook on payment error.
Should payment fail for whatever reason as well as throw an exception a
hook is added so that third parties can run any routines they need to in
this scenario.

Use case for this change is we want to be able to log on any payment
error (to new relic).

Stripe has a similar hook here: https://github.com/woocommerce/woocommerce-gateway-stripe/blob/master/includes/class-wc-stripe-order-handler.php#L191