Exit early from scripts when no form is present
Currently, when on a product page containing an item that is out of
stock, both wc-gateway-ppec-smart-payment-buttons.js and
wc-gateway-ppec-generate-cart.js scripts throw exceptions as both
scripts are trying to target elements which don't exist when products
are out of stock.

This PR adds a check for the relevant form element, and returns early if
it is not present.

Fixes #558